move exceptions into a file of their own
